Moved RunInfo class to wrfcloud.runtime.__init__ to resolve the circular dependency.
************* Module wrfcloud.runtime.tools.make_wps_namelist
wrfcloud/runtime/tools/make_wps_namelist.py:1:0: R0401: Cyclic import (wrfcloud.runtime.real -> wrfcloud.runtime.run) (cyclic-import)
wrfcloud/runtime/tools/make_wps_namelist.py:1:0: R0401: Cyclic import (wrfcloud.runtime.metgrid -> wrfcloud.runtime.tools.make_wps_namelist -> wrfcloud.runtime.run) (cyclic-import)
wrfcloud/runtime/tools/make_wps_namelist.py:1:0: R0401: Cyclic import (wrfcloud.runtime.run -> wrfcloud.runtime.ungrib -> wrfcloud.runtime.tools.make_wps_namelist) (cyclic-import)
wrfcloud/runtime/tools/make_wps_namelist.py:1:0: R0401: Cyclic import (wrfcloud.runtime.postproc -> wrfcloud.runtime.run) (cyclic-import)
wrfcloud/runtime/tools/make_wps_namelist.py:1:0: R0401: Cyclic import (wrfcloud.runtime.run -> wrfcloud.runtime.wrf) (cyclic-import)
wrfcloud/runtime/tools/make_wps_namelist.py:1:0: R0401: Cyclic import (wrfcloud.runtime.run -> wrfcloud.runtime.ungrib) (cyclic-import)
