+++ python/src/wrfcloud/runtime/real.py
@@ -6,7 +6,7 @@ Functions for setting up, executing, and monitoring a run of the 'real.exe' WRF 
 from logging import Logger
 
 # Import our custom modules
-from wrfcloud.runtime.run import RunInfo
+from wrfcloud.runtime import RunInfo
 
 
 def main(runinfo: RunInfo, logger: Logger) -> None:
