+++ python/src/wrfcloud/runtime/run.py
@@ -11,7 +11,6 @@ run.
 import argparse
 import logging
 import os
-import yaml
 
 # Import our custom modules
 from wrfcloud.runtime import ungrib
@@ -19,42 +18,7 @@ from wrfcloud.runtime import metgrid
 from wrfcloud.runtime import real
 from wrfcloud.runtime import wrf
 from wrfcloud.runtime import postproc
-
-# Define our Classes
-class RunInfo:
-    """
-    This class keeps info about the run
-    """
-
-    def __init__(self, name: str):
-        self.name = name
-        self.topdir = os.getcwd()
-        self.wd = self.topdir + '/' + name
-        logging.debug(f'Working directory set to {self.wd}')
-        self.staticdir = self.topdir + '/domains/' + name
-        logging.debug(f'Static data directory is {self.staticdir}')
-        self.read_config(name)
-
-        self.ungribdir = self.wd + '/ungrib'
-        self.metgriddir = self.wd + '/metgrid'
-
-    def read_config(self, name: str) -> None:
-        """
-        This method reads the config file for this run, and sets the appropriate variables
-        for this class.
-        """
-        config_file=name + '.yml'
-        logging.debug(f'reading config file {config_file}')
-        with open(config_file, 'r', encoding='utf-8') as file:
-            config = yaml.safe_load(file)
-        logging.debug(f'Read {config_file} successfully, values are:')
-        logging.debug(f'{config}')
-        self.configuration = config['run']['configuration']
-        self.startdate = config['run']['start']
-        self.enddate = config['run']['end']
-        self.input_freq_sec = config['run']['output_freq_sec']
-        self.output_freq_sec = config['run']['output_freq_sec']
-        self.local_data = config['run']['local_data']
+from wrfcloud.runtime import RunInfo
 
 
 # Define our functions
